Separate the new logic
